More Frontend Cleanup
