+++ frontend/src/components/collaborative/Collaborative.module.scss
@@ -1,0 +1,3 @@
+.collaborativeWrapper {
+    position: relative;
+}
