--- frontend/src/components/collaborative/styles/Collaborative.module.scss
@@ -1,16 +1,0 @@
-.collaborativeWrapper {
-    position: relative;
-
-    .winScreen {
-        position: absolute;
-        border: 1px solid black;
-        top: 100px;
-        left: 0;
-        right: 0;
-        width: 400px;
-        margin-left: auto;
-        margin-right: auto;
-        background-color: white;
-        padding: 20px;
-    }
-}
