+++ frontend/src/components/competitive/Competitive.module.scss
