+++ frontend/src/components/header/Header.module.scss
@@ -1,0 +1,12 @@
+.winScreen {
+    position: absolute;
+    border: 1px solid black;
+    top: 100px;
+    left: 0;
+    right: 0;
+    width: 400px;
+    margin-left: auto;
+    margin-right: auto;
+    background-color: white;
+    padding: 20px;
+}
