+++ frontend/src/components/header/Paused.tsx
